+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Command predicates are now evaluated lazily. This means that you can create
+  predicates which depend on previous predicates to evaluate as `:passthrough`,
+  for example, a permission predicate might want a `guild_only` predicate used
+  before.
 
 ## 0.1.0 - 03.02.2019
 Initial release.